move adapters/ to stdout_adapter/.
