+++ docs/cookbook.rst
@@ -267,8 +267,10 @@ Here's one readable way to deal with all these ``config`` objects::
 
     import screenpy
     import screenpy_some_extension
+    from screenpy.narration import stdout_adapter
 
     screenpy.config.TIMEOUT = 42
+    stdout_adapter.config.INDENT_CHAR = ">"
     screenpy_some_extension.config.SOME_SETTING = "spam"
 
 That way,
