+++ docs/settings.rst
@@ -51,5 +51,5 @@ ScreenPy Default Settings
 ScreenPy StdOutAdapter Default Settings
 ---------------------------------------
 
-.. autopydantic_settings:: screenpy.narration.adapters.settings.StdOutAdapterSettings
+.. autopydantic_settings:: screenpy.narration.stdout_adapter.settings.StdOutAdapterSettings
     :exclude-members: Config.customise_sources
