+++ screenpy/narration/__init__.py
@@ -2,9 +2,9 @@
 Tools for narrating the screenplay.
 """
 
-from .adapters import LoggingAdapter, StdOutAdapter, StdOutManager
 from .gravitas import AIRY, EXTREME, HEAVY, LIGHT, NORMAL
 from .narrator import Narrator
+from .stdout_adapter import LoggingAdapter, StdOutAdapter, StdOutManager
 
 __all__ = [
     "AIRY",
