+++ screenpy/narration/stdout_adapter/__init__.py
