+++ screenpy/narration/stdout_adapter/stdout_adapter.py
