+++ tests/test_namespace.py
@@ -266,11 +266,11 @@ def test_narration():
     assert sorted(screenpy.narration.__all__) == sorted(expected)
 
 
-def test_adapters():
+def test_stdout_adapter():
     expected = [
         "config",
         "LoggingAdapter",
         "StdOutAdapter",
         "StdOutManager",
     ]
-    assert sorted(screenpy.narration.adapters.__all__) == sorted(expected)
+    assert sorted(screenpy.narration.stdout_adapter.__all__) == sorted(expected)
